move model settings button in ManageModel area
